daemon: C's size_t reads, accept() log lines and helper reuse
cpu cores, the malloc arena maxes and web server max sockets are read into a
size_t as in C (a negative value wraps), and the libuv worker count is
computed in int from the truncated CPU count; replication threads take
(int)MAX(cpus / 3, 4). An ACL-denied connection logs C's second line
(accept() failed, EPERM) and the EMFILE line is limited to one per 10 s.
PATH and PYTHONPATH keep non-UTF-8 values, PATH is cut at 4095 bytes, and a
missing plugins directory prints (null) as C does. system.rs uses the text
crate's str2uint64 and find (skipping leading spaces like C's str2ull), and
the malloc export reuses conf::export. Stale comments corrected.

+++ src/crates/netdata-agent/streaming/src/conf.rs
@@ -33,8 +33,8 @@ const KEEPALIVE_IDLE_MAX_SECONDS: u64 = 3600;
 /// Values `stream_conf_load()` takes from the rest of the daemon.
 #[derive(Debug, Clone, Copy)]
 pub struct LoadDefaults {
-    /// `netdata_conf_cpus()`, for `replication_threads_default()`.
-    pub conf_cpus: i64,
+    /// `netdata_conf_cpus()` (a `size_t`), for `replication_threads_default()`.
+    pub conf_cpus: u64,
     /// `libuv_worker_threads`, for `replication_prefetch_default()`.
     pub libuv_worker_threads: i64,
     /// `netdata_ssl_validate_certificate`.
@@ -240,7 +240,8 @@ impl StreamConf {
         let has_api_enabled = self.config.stream_conf_has_api_enabled();
         let parent_profile = is_parent_profile(netdata, has_api_enabled, self.send.enabled);
         let threads_default = if parent_profile {
-            (defaults.conf_cpus / 3).max(4)
+            // (int)MAX(cpus / 3, 4)
+            i64::from((defaults.conf_cpus / 3).max(4) as i32)
         } else {
             1
         };
